[num-variants] rename to enum-macros
This can potentially be used for other enum-related macros soon.
